Add --[no-]pl011 option
MacOS VMs emit their debug log to a PL011 serial port, not the
virtio-console one that we spawn with --serial. This is important if you
want to see kernel debug output. Let's add an option to turn on the pl011
console as well. For now, it can only be attached to stdin/stdout.

Signed-off-by: Alexander Graf <graf@amazon.com>