Add support for multiple automount shares
The Virtualization framework support having multiple automounted shares with different readonly vs readwrite statuses. But all of the automount shares have to be specified as part of a single VZVirtioFileSystemDeviceConfiguration.
The current implementation of macosvm doesn't do that, as it would address each automount share individually, which means the first one would be allowed and the second would result in an error.
> validateWithError = Error Domain=VZErrorDomain Code=2 "More than one Virtio File System device uses the tag “com.apple.virtio-fs.automount”." UserInfo={NSLocalizedFailure=Invalid virtual machine configuration., NSLocalizedFailureReason=More than one Virtio File System device uses the tag “com.apple.virtio-fs.automount”.}

There is already a starting point for this with VMInstance.m's "addAutomountDirectoryShares", but it is unused, and would force that every automounted share must have the same readonly vs readwrite status.
So to fix this, I've split up the share configuration logic into two loops.
First loop over all shares. If there are any automount shares, add them to a second list and move on. Otherwise configure the share as normal.
Second loop is just over the automount shares that were found. It batches all of the automount directories into a single VZVirtioFileSystemDeviceConfiguration, which allows them to all be registered without error.